Responding to PR feedback
