+++ lale/lib/sklearn/ada_boost_classifier.py
@@ -18,29 +18,15 @@ from sklearn.ensemble import AdaBoostClassifier as SKLModel
 import lale.docstrings
 import lale.operators
 
+from .fit_spec_proxy import _FitSpecProxy
 from .function_transformer import FunctionTransformer
-
-
-class FitSpecProxy:
-    def __init__(self, base):
-        self._base = base
-
-    def __getattr__(self, item):
-        return getattr(self._base, item)
-
-    def get_params(self, deep=True):
-        ret = {}
-        ret["base"] = self._base
-        return ret
-
-    def fit(self, X, y, sample_weight=None, **fit_params):
-        return self._base.fit(X, y, sample_weight=sample_weight, **fit_params)
 
 
 class _AdaBoostClassifierImpl:
     def __init__(
         self,
         base_estimator=None,
+        *,
         n_estimators=50,
         learning_rate=1.0,
         algorithm="SAMME.R",
@@ -49,7 +35,7 @@ class _AdaBoostClassifierImpl:
         if base_estimator is None:
             estimator_impl = None
         else:
-            estimator_impl = FitSpecProxy(base_estimator)
+            estimator_impl = _FitSpecProxy(base_estimator)
 
         self._hyperparams = {
             "base_estimator": estimator_impl,
@@ -74,7 +60,7 @@ class _AdaBoostClassifierImpl:
                 inverse_func=None,
                 check_inverse=False,
             )
-            self._hyperparams["base_estimator"] = FitSpecProxy(
+            self._hyperparams["base_estimator"] = _FitSpecProxy(
                 feature_transformer >> self._hyperparams["base_estimator"]
             )
             self._wrapped_model = SKLModel(**self._hyperparams)
