+++ lale/lib/sklearn/ada_boost_regressor.py
@@ -17,8 +17,8 @@ from sklearn.ensemble import AdaBoostRegressor as SKLModel
 
 import lale.docstrings
 import lale.operators
-from lale.lib.sklearn.ada_boost_classifier import FitSpecProxy
 
+from .fit_spec_proxy import _FitSpecProxy
 from .function_transformer import FunctionTransformer
 
 
@@ -26,6 +26,7 @@ class _AdaBoostRegressorImpl:
     def __init__(
         self,
         base_estimator=None,
+        *,
         n_estimators=50,
         learning_rate=1.0,
         loss="linear",
@@ -34,7 +35,7 @@ class _AdaBoostRegressorImpl:
         if base_estimator is None:
             estimator_impl = None
         else:
-            estimator_impl = FitSpecProxy(base_estimator)
+            estimator_impl = _FitSpecProxy(base_estimator)
 
         self._hyperparams = {
             "base_estimator": estimator_impl,
@@ -59,7 +60,7 @@ class _AdaBoostRegressorImpl:
                 inverse_func=None,
                 check_inverse=False,
             )
-            self._hyperparams["base_estimator"] = FitSpecProxy(
+            self._hyperparams["base_estimator"] = _FitSpecProxy(
                 feature_transformer >> self._hyperparams["base_estimator"]
             )
             self._wrapped_model = SKLModel(**self._hyperparams)
